// src/api/openapi/model/offerTemplateListElement.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OfferTemplateListElement { 
    description: string;
    entry_count: number;
    id: number;
    name: string;
}